[#3] 프로젝트 초기설정 (#5)
* feat : 첫화면 추가 로그인->로그인, 시작하기->회원가입

* ci : move project-structre-depth

* ci : add yarn.lock

* ci : add .gitignore windows,mac,vscode,nextjs

* ci : add .gitignore macos, windows, nextjs, react

* fix : fix .gitignore

* style : add eslint, prettier

* style : eslint error 관련 코드 수정

---------
