+++ .vscode/settings.json
@@ -1,0 +1,18 @@
+{
+  "editor.formatOnSave": true,
+  "editor.codeActionsOnSave": {
+    // For ESLint
+    "source.fixAll.eslint": true
+  },
+  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  // // Enable per-language
+  // "[typescript]": {
+  //   // "editor.formatOnSave": true,
+  //   "editor.defaultFormatter": "esbenp.prettier-vscode"
+  // },
+  // "[json]": {
+  //   // "editor.formatOnSave": true
+  //   "editor.defaultFormatter": "esbenp.prettier-vscode"
+  // },
+  "workbench.settings.useSplitJSON": true
+}
