--- front-end/pages/main/index.tsx
@@ -1,9 +1,0 @@
-function Main() {
-  return (
-    <div>
-      Main
-    </div>
-  )
-}
-
-export default Main;
