--- front-end/pages/signup.tsx
@@ -1,7 +1,0 @@
-export default function SignUp() {
-  return (
-    <div>
-      회원가입
-    </div>
-  )
-}
