+++ next.config.js
@@ -1,10 +1,10 @@
 /** @type {import('next').NextConfig} */
-const withPWA = require("next-pwa")({
-    dest: "public",
+const withPWA = require('next-pwa')({
+  dest: 'public',
 });
 
 const nextConfig = {
   // reactStrictMode: true,
-}
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
