+++ package.json
@@ -13,7 +13,8 @@
     "@types/react": "18.0.38",
     "@types/react-dom": "18.0.11",
     "autoprefixer": "10.4.14",
-    "eslint": "8.39.0",
+    "eslint": "^8.39.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-next": "13.3.1",
     "next": "13.3.1",
     "next-pwa": "^5.6.0",
@@ -22,5 +23,12 @@
     "react-dom": "18.2.0",
     "tailwindcss": "3.3.1",
     "typescript": "5.0.4"
+  },
+  "devDependencies": {
+    "@babel/eslint-parser": "^7.21.3",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "^2.8.8"
   }
 }
