+++ pages/main/index.tsx
@@ -1,0 +1,5 @@
+function Main() {
+  return <div>Main</div>;
+}
+
+export default Main;
