+++ pages/signup.tsx
@@ -1,0 +1,3 @@
+export default function SignUp() {
+  return <div>회원가입</div>;
+}
